requirements.txt
requests==2.32.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
fastapi==0.111.1